Add files via upload
Started working locally. With this current status, you can choose whichever Pokemon to fight. The rival is chosen randomly. It uses PokeAPI. You can choose the level as well and movements and stats are calculated appropiately. You can choose which move to use and moves that are not status type and have power already cause damage.